Rename tests to use with the failsafe plugin
