+++ appserver/extras/embedded/tests/pom.xml
@@ -59,7 +59,7 @@
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-surefire-plugin</artifactId>
+                <artifactId>maven-failsafe-plugin</artifactId>
                 <configuration>
                     <reuseForks>false</reuseForks>
                 </configuration>
