+++ appserver/extras/embedded/tests/src/test/java/org/glassfish/main/extras/embedded/test/all/deployment/ejb/RemoteEjbJarDeploymentITest.java
@@ -14,7 +14,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package org.glassfish.main.extras.embedded.test.all.deployment.webapp;
+package org.glassfish.main.extras.embedded.test.all.deployment.ejb;
 
 import java.io.File;
 import java.io.IOException;
@@ -24,12 +24,11 @@ import javax.naming.NamingException;
 
 import org.glassfish.embeddable.Deployer;
 import org.glassfish.embeddable.GlassFishException;
-import org.glassfish.main.extras.embedded.test.all.deployment.RemoteDeploymentTestBase;
+import org.glassfish.main.extras.embedded.test.all.deployment.RemoteDeploymentITestBase;
 import org.glassfish.main.extras.embedded.test.app.ejb.RemoteBean;
 import org.glassfish.main.extras.embedded.test.app.ejb.RemoteInterface;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.JavaArchive;
-import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.junit.jupiter.api.Test;
 
 import static java.lang.System.Logger.Level.INFO;
@@ -37,41 +36,36 @@ import static java.lang.System.Logger.Level.WARNING;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
 
-public class RemoteWebAppDeploymentTest extends RemoteDeploymentTestBase {
+public class RemoteEjbJarDeploymentITest extends RemoteDeploymentITestBase {
 
-    private static final System.Logger LOG = System.getLogger(RemoteWebAppDeploymentTest.class.getName());
+    private static final System.Logger LOG = System.getLogger(RemoteEjbJarDeploymentITest.class.getName());
 
-    private static final String APP_NAME = "RemoteWebApp";
-
-    private static final String LIB_FILE_NAME = "RemoteLib.jar";
+    private static final String APP_NAME = "RemoteEjbJar";
 
     @Test
     public void testDeployAndUndeployApplication() throws GlassFishException, IOException, NamingException {
         Deployer deployer = glassfish.getDeployer();
-        File warFile = createDeployment();
+        File ejbFile = createDeployment();
         try {
-            assertThat(deployer.deploy(warFile), equalTo(APP_NAME));
+            assertThat(deployer.deploy(ejbFile), equalTo(APP_NAME));
             assertThat(getResult(APP_NAME), equalTo(Runtime.version().toString()));
         } finally {
             try {
-                Files.deleteIfExists(warFile.toPath());
-            } catch (IOException e){
-                LOG.log(WARNING, "An error occurred while remove temp file " + warFile.getAbsolutePath(), e);
+                Files.deleteIfExists(ejbFile.toPath());
+            } catch (IOException e) {
+                LOG.log(WARNING, "An error occurred while remove temp file " + ejbFile.getAbsolutePath(), e);
             }
             deployer.undeploy(APP_NAME);
         }
     }
 
     private File createDeployment() throws IOException {
-        JavaArchive javaArchive = ShrinkWrap.create(JavaArchive.class, LIB_FILE_NAME)
-            .addClass(RemoteInterface.class);
+        JavaArchive ejbArchive = ShrinkWrap.create(JavaArchive.class)
+            .addClass(RemoteInterface.class)
+            .addClass(RemoteBean.class);
 
-        WebArchive webArchive = ShrinkWrap.create(WebArchive.class)
-            .addClass(RemoteBean.class)
-            .addAsLibrary(javaArchive);
+        LOG.log(INFO, ejbArchive.toString(true));
 
-        LOG.log(INFO, webArchive.toString(true));
-
-        return createFileFor(webArchive, APP_NAME);
+        return createFileFor(ejbArchive, APP_NAME);
     }
 }
